moved functions to handle thresholds to string_manip utility
